refactor: extract OrderFillsAt component
